requirements.txt
# Core
openai
python-dotenv
requests
tqdm

# PDF
PyMuPDF

# Data / ML
numpy
pandas
torch

# NLP (sentence splitting)
spacy